+++ src/scripts/index.js
@@ -1,3 +1,9 @@
 import projects from "./projectsData.js";
+import printToDom from "./printToDom.js";
+import createProjectCards from "./createProjectCards.js";
 
-export {projects};
+export {
+  projects,
+  printToDom,
+  createProjectCards,
+};
