+++ src/scripts/init.js
@@ -1,0 +1,7 @@
+
+import {createProjectCards, projects} from "./index.js";
+
+const init = () => {
+  createProjectCards(projects);
+};
+init();
