+++ src/scripts/printToDom.js
@@ -1,0 +1,7 @@
+
+const printToDom = (divId, textToPrint) => {
+  const selectedDiv = document.querySelector(divId);
+  selectedDiv.innerHTML = textToPrint;
+};
+
+export default printToDom;
